NODEDynamics

Documentation for NODEDynamics.

NODEDynamics.NODE

NODEDynamics.NODE

NODEDynamics.accuracy

NODEDynamics.modelEnv

NODEDynamics.modelEnv

NODEDynamics.train_step!

NODEDynamics.train_step!

NODEDynamics.NODE — Method

(m::NODE)(timestamps, datapoints)

Apply the NODE model to the given timestamps and datapoints.

NODEDynamics.NODE — Method

NODE(input_size::Int, ode_size::Int, output_size::Int)

Create a NODE model with the given input, ode and output size.

NODEDynamics.accuracy — Method

accuracy(y_true, y_pred, tolerance)

Calculate the accuracy of the prediction given the true values and a tolerance.

NODEDynamics.modelEnv — Method

modelEnv(environment::ContinuousEnvironment, modelParams::ModelParameter)

Main algorithm to train the continuous action model using the given environment and model parameters.

NODEDynamics.modelEnv — Method

modelEnv(environment::DiscreteEnvironment, modelParams::ModelParameter)

Main algorithm to train the discrete action model using the given environment and model parameters.

NODEDynamics.train_step! — Method

train_step!(environment::ContinuousEnvironment, S, A, R, S´, T, fθ, model_opt)

Train the continuous action model using the given data with gradient descent.

NODEDynamics.train_step! — Method

train_step!(environment::DiscreteEnvironment, S, A, R, S´, T, fθ, model_opt, ep::EnvParameter)

Train the discrete action model using the given data with gradient descent.