NODEDynamics

Documentation for NODEDynamics.

NODEDynamics.jl learns a model of the environment. It works with the following environments:

Pendulum

Acrobot

Lunar Lander

Installation

In the julia REPL, run

using Pkg
Pkg.add(url="https://github.com/SvenDuve/NODEDynamics.jl")

This package is not for general use. To be fully functional, it requires the following packages:

RLTypes

Conda

PyCall

Gymnasium, this is a slightly adapted version. The original package can be found here

Run the following:

using Pkg
Pkg.add(url="https://github.com/SvenDuve/RLTypes.jl")
Pkg.add("Conda")
Pkg.add("PyCall")

Then again within julia,

using Conda
using PyCall
Conda.pip("install", "git+https://github.com/SvenDuve/Gymnasium")

Bring package into scope with

using NODEDynamics

Example usage

Once the above is installed, the following code can be used to train an agent on data collected form 100 episodes of the Pendulum environment:

using NODEDynamics
using RLTypes
model(Pendulum(), ModelParameter(collect_train=100, collect_test=10))

For further parameters, see the code.

This function returns a data structure containing a trained model.

NODEDynamics.NODE

NODEDynamics.NODE

NODEDynamics.accuracy

NODEDynamics.modelEnv

NODEDynamics.modelEnv

NODEDynamics.train_step!

NODEDynamics.train_step!

NODEDynamics.NODE — Method

(m::NODE)(timestamps, datapoints)

Apply the NODE model to the given timestamps and datapoints.

NODEDynamics.NODE — Method

NODE(input_size::Int, ode_size::Int, output_size::Int)

Create a NODE model with the given input, ode and output size.

NODEDynamics.accuracy — Method

accuracy(y_true, y_pred, tolerance)

Calculate the accuracy of the prediction given the true values and a tolerance.

NODEDynamics.modelEnv — Method

modelEnv(environment::ContinuousEnvironment, modelParams::ModelParameter)

Main algorithm to train the continuous action model using the given environment and model parameters.

NODEDynamics.modelEnv — Method

modelEnv(environment::DiscreteEnvironment, modelParams::ModelParameter)

Main algorithm to train the discrete action model using the given environment and model parameters.

NODEDynamics.train_step! — Method

train_step!(environment::ContinuousEnvironment, S, A, R, S´, T, fθ, model_opt)

Train the continuous action model using the given data with gradient descent.

NODEDynamics.train_step! — Method

train_step!(environment::DiscreteEnvironment, S, A, R, S´, T, fθ, model_opt, ep::EnvParameter)

Train the discrete action model using the given data with gradient descent.